refactor:  how tags are stored
remove strapi as dependency and just store tags as a plain ts file
