+++ src/commands/support/tag.ts
@@ -1,8 +1,8 @@
-import { Command, CommandoClient, CommandoMessage } from 'discord.js-commando';
 import { Message, MessageEmbed } from 'discord.js';
-import getTags from '../../tagsCache';
-import { Tag as ITag } from '../../tagsCache';
+import { Command, CommandoClient, CommandoMessage } from 'discord.js-commando';
+
 import getCommands from '../../commandsCache';
+import { Tag as ITag, tags } from '../../tags';
 
 
 export default class Tag extends Command {
@@ -17,7 +17,6 @@ export default class Tag extends Command {
     }
 
     async run(message: CommandoMessage, args: string): Promise<Message | Message[]> {
-        const tags = await getTags();
         const tagToSend = await determineReply(args);
 
         if (!tagToSend) {
@@ -34,8 +33,6 @@ export default class Tag extends Command {
 }
 
 async function determineReply(textToSearch): Promise<ITag> {
-    // First try to match with a tag
-    const tags = await getTags();
 
     const tagToSend = tags.filter(t => t.code).find(t => t.code.toLowerCase() === textToSearch.toLowerCase());
 
@@ -59,7 +56,6 @@ async function determineReply(textToSearch): Promise<ITag> {
 
             ${commandToSend.extendedDescription}`,
             code: '',
-            trigger: ''
         }
 
         return data;
