+++ test/teardown.js
@@ -1,4 +1,3 @@
 module.exports = async function () {
-    await global.gqlServer.close();
     await global.docsServer.close();
 };
